Don't cast IB_NODE_VENDORID_F, IB_NODE_DEVID_F to uint16.
IB_NODE_VENDORID_F is 24 bits, whilst IB_NODE_DEVID_F is 16 bits. This commit
trusts mad_get_field() to return the right number of bits, and stores the
value in a uint (guaranteed to be at least 32 bits).

+++ infiniband/infiniband.go
@@ -29,8 +29,8 @@ type Node struct {
 	GUID     uint64
 	NodeType int
 	NodeDesc string
-	VendorID uint16
-	DeviceID uint16
+	VendorID uint
+	DeviceID uint
 	Ports    []Port
 }
